Вынесла в отдельный модель функцию
