+++ js/main.js
@@ -1,1 +1,1 @@
-import './data.js';
+import './offer.js';
